Moved API calls into their own class
