+++ app/representers/openstax/accounts/api/v1/group_nesting_representer.rb
@@ -1,6 +1,6 @@
 module OpenStax
   module Accounts
-    module Api
+    class Api
       module V1
         class GroupNestingRepresenter < Roar::Decorator
           include Roar::Representer::JSON
