+++ lib/openstax/accounts/version.rb
@@ -1,5 +1,5 @@
 module OpenStax
   module Accounts
-    VERSION = "4.1.1"
+    VERSION = "5.0.0"
   end
 end
